ADEN-13135-fix-optimizely-dw (#1987)
