+++ src/platforms/f2/f2-platform.ts
@@ -30,6 +30,7 @@ import { F2BaseContextSetup } from './setup/context/base/f2-base-context.setup';
 import { F2SlotsContextSetup } from './setup/context/slots/f2-slots-context.setup';
 import { F2TargetingSetup } from './setup/context/targeting/f2-targeting.setup';
 import { F2DynamicSlotsSetup } from './setup/dynamic-slots/f2-dynamic-slots.setup';
+import { F2ExperimentsSetup } from './setup/experiments/f2-experiments.setup';
 import { F2TemplatesSetup } from './templates/f2-templates.setup';
 
 @Injectable()
@@ -55,6 +56,7 @@ export class F2Platform {
 			F2TemplatesSetup,
 			SequentialMessagingSetup, // SequentialMessagingSetup needs to be after *TemplatesSetup or UAP SM might break
 			LabradorSetup,
+			F2ExperimentsSetup,
 			TrackingSetup,
 			AdEngineRunnerSetup,
 			() => communicationService.emit(eventsRepository.AD_ENGINE_CONFIGURED),
