+++ src/platforms/f2/templates/handlers/bfaa/bfaa-f2-config-handler.ts
@@ -3,7 +3,6 @@ import {
 	communicationService,
 	context,
 	eventsRepository,
-	Optimizely,
 	TEMPLATE,
 	TemplateStateHandler,
 	UapParams,
@@ -11,22 +10,11 @@ import {
 } from '@wikia/ad-engine';
 import { Inject, Injectable } from '@wikia/dependency-injection';
 
-const OPTIMIZELY_NTC_2_0 = {
-	EXPERIMENT_ENABLED: 'unified_takeover_ntc20',
-	EXPERIMENT_VARIANT: 'unified_takeover_ntc20_variant',
-};
-
-const OPTIMIZELY_NTC_2_0_VARIANTS = {
-	NEW_VARIANT: 'ntc20_adhesion_enabled',
-	OLD_VARIANT: 'ntc20_adhesion_disabled',
-	UNDEFINED: 'ntc20_adhesion_undefined',
-};
-
 @Injectable({ autobind: false })
 export class BfaaF2ConfigHandler implements TemplateStateHandler {
 	private enabledSlots: string[] = ['top_boxad', 'incontent_boxad', 'bottom_leaderboard'];
 
-	constructor(@Inject(TEMPLATE.PARAMS) private params: UapParams, private optimizely: Optimizely) {}
+	constructor(@Inject(TEMPLATE.PARAMS) private params: UapParams) {}
 
 	async onEnter(): Promise<void> {
 		const isMobile = context.get('state.isMobile');
@@ -60,21 +48,8 @@ export class BfaaF2ConfigHandler implements TemplateStateHandler {
 	}
 
 	private configureFloorAdhesionExperiment() {
-		this.optimizely.addVariantToTargeting(
-			OPTIMIZELY_NTC_2_0,
-			OPTIMIZELY_NTC_2_0_VARIANTS.UNDEFINED,
-		);
-
-		const variant = this.optimizely.getVariant(OPTIMIZELY_NTC_2_0);
-
-		if (variant) {
-			this.optimizely.addVariantToTargeting(OPTIMIZELY_NTC_2_0, variant);
-
-			if (variant === OPTIMIZELY_NTC_2_0_VARIANTS.NEW_VARIANT) {
-				this.enabledSlots.push('floor_adhesion');
-
-				document.body.classList.add('floor-adhesion-experiment');
-			}
+		if (context.get('options.ntc.floorEnabled')) {
+			this.enabledSlots.push('floor_adhesion');
 		}
 	}
 }
